Create Advertisement_Click_Prediction
Advertisement click prediction is the process of using machine learning algorithms to predict whether or not a user will click on an advertisement. This information can be used to improve the effectiveness of online advertising campaigns by targeting ads to users who are more likely to be interested in them.