Advertisement_Click_Prediction
Machine Learning has now variety of applications in real world whether it is for predictive analysis or automating things or decision making for business purpose. Having knowledge of Machine Learning is different thing but applying it to business purpose for making profit, and learning something new from the data is the main motive. Project was about predicting the ad click rate of a customer, based on certain features. I found this problem very interesting and therefore made a Machine Learning model to predict the Ad click.